Added logout button for logging out the user session
